+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- CI: `pre-commit autoupdate`
+
+### Fixed
+
+- `Solution`: `get_total_amount` and related methods could fail when the oxidation state of an element was
+  unknown (e.g., 'Br') (Issue [#116](https://github.com/KingsburyLab/pyEQL/issues/116))
+
 ## [0.14.0] - 2024-03-05
 
 ### Added